pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a22a"
version = "0.1.0"
description = "All-22 Atlas bootstrap scaffolding"
authors = [{name = "All-22 Atlas"}]
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "pyyaml>=6.0",
    "polars>=0.20",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["."]
